Add search function. Information of visited documents will be recorded in database
To display a document, we need two arguments: document id and its page number. So we need information of documents, especially page number. When a document is requested, first query its information in database. If found, render the document. If not, send a request to docin.com to get its information and then save information in the database. If there is no such a document on docin.com, attach invalid bit to this id in database.

   TODO issue: Some old documents on doc.com don't have document information. Need a better way to get its page number.